Moved request/response handling to separated set
