+++ crates/world_server/src/set/mod.rs
@@ -2,14 +2,16 @@
 
 mod chunk_initialization;
 mod chunk_management;
-// mod collect_dispatch;
 mod landscape;
 mod meshing;
 mod propagation;
+mod receive_requests;
+mod send_responses;
 
 pub use chunk_initialization::*;
 pub use chunk_management::*;
-// pub use collect_dispatch::*;
 pub use landscape::*;
 pub use meshing::*;
 pub use propagation::*;
+pub(crate) use receive_requests::*;
+pub(crate) use send_responses::*;
